# moviesapiswagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Movies API",
    "version": "v1"
  },
  "paths": {
    "/api/films": {
      "get": {
        "tags": [
          "Films"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/movies": {
      "get": {
        "tags": [
          "Movies"
        ],
        "operationId": "GetMovies",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              },
              "text/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Movies"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieForCreation"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieForCreation"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MovieForCreation"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/MovieForCreation"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/MovieForCreation"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/MovieForCreation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/movies/{movieId}": {
      "get": {
        "tags": [
          "Movies"
        ],
        "operationId": "GetMovie",
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieForUpdate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieForUpdate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MovieForUpdate"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/MovieForUpdate"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/MovieForUpdate"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/MovieForUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieForUpdateJsonPatchDocument"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieForUpdateJsonPatchDocument"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MovieForUpdateJsonPatchDocument"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/movies/{movieid}": {
      "delete": {
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/moviesstream": {
      "get": {
        "tags": [
          "MoviesStream"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/movies/{movieId}/posters/{posterId}": {
      "get": {
        "tags": [
          "Posters"
        ],
        "operationId": "GetPoster",
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "posterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Poster"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Poster"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Poster"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Poster"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Poster"
                }
              }
            }
          }
        }
      }
    },
    "/api/movies/{movieId}/posters": {
      "post": {
        "tags": [
          "Posters"
        ],
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PosterForCreation"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PosterForCreation"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PosterForCreation"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/PosterForCreation"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/PosterForCreation"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/PosterForCreation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/movies/{movieId}/trailers/{trailerId}": {
      "get": {
        "tags": [
          "Trailers"
        ],
        "operationId": "GetTrailer",
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "trailerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Trailer"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trailer"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trailer"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Trailer"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Trailer"
                }
              }
            }
          }
        }
      }
    },
    "/api/movies/{movieId}/trailers": {
      "post": {
        "tags": [
          "Trailers"
        ],
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrailerForCreation"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TrailerForCreation"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TrailerForCreation"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/TrailerForCreation"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/TrailerForCreation"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/TrailerForCreation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "IContractResolver": {
        "type": "object",
        "additionalProperties": false
      },
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "genre": {
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "type": "string",
            "format": "date-time"
          },
          "director": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MovieForCreation": {
        "required": [
          "directorId",
          "releaseDate",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 200,
            "type": "string"
          },
          "description": {
            "maxLength": 2000,
            "minLength": 10,
            "type": "string",
            "nullable": true
          },
          "genre": {
            "maxLength": 200,
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "type": "string",
            "format": "date-time"
          },
          "directorId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "MovieForUpdate": {
        "required": [
          "directorId",
          "releaseDate",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 200,
            "type": "string"
          },
          "description": {
            "maxLength": 2000,
            "type": "string",
            "nullable": true
          },
          "genre": {
            "maxLength": 200,
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "type": "string",
            "format": "date-time"
          },
          "directorId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "MovieForUpdateJsonPatchDocument": {
        "type": "object",
        "properties": {
          "operations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovieForUpdateOperation"
            },
            "nullable": true
          },
          "contractResolver": {
            "$ref": "#/components/schemas/IContractResolver"
          }
        },
        "additionalProperties": false
      },
      "MovieForUpdateOperation": {
        "type": "object",
        "properties": {
          "operationType": {
            "$ref": "#/components/schemas/OperationType"
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "op": {
            "type": "string",
            "nullable": true
          },
          "from": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OperationType": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
      "Poster": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "movieId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "bytes": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PosterForCreation": {
        "required": [
          "bytes",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 200,
            "type": "string"
          },
          "bytes": {
            "type": "string",
            "format": "byte"
          }
        },
        "additionalProperties": false
      },
      "Trailer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "movieId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "bytes": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TrailerForCreation": {
        "required": [
          "bytes",
          "movieId",
          "name"
        ],
        "type": "object",
        "properties": {
          "movieId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 200,
            "type": "string"
          },
          "description": {
            "maxLength": 1000,
            "type": "string",
            "nullable": true
          },
          "bytes": {
            "type": "string",
            "format": "byte"
          }
        },
        "additionalProperties": false
      }
    }
  }
}